% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{This is a function to check for leap years}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{The year to check}
}
\value{
TRUE if it's a leap year, FALSE otherwise
}
\description{
This is a function to check for leap years
}
